requirements.txt
streamlit
pillow
fpdf
openai
requests
